+++ Mods/SOD.LifeAndLiving/Plugin.cs
@@ -15,14 +15,14 @@ namespace SOD.LifeAndLiving
     {
         public const string PLUGIN_GUID = "Venomaus.SOD.LifeAndLiving";
         public const string PLUGIN_NAME = "LifeAndLiving";
-        public const string PLUGIN_VERSION = "1.0.3";
+        public const string PLUGIN_VERSION = "1.0.4";
 
         public override void Load()
         {
             Lib.SaveGame.OnBeforeLoad += SaveGame_OnBeforeLoad;
             Lib.SaveGame.OnBeforeSave += SaveGame_OnBeforeSave;
             Lib.SaveGame.OnBeforeDelete += SaveGame_OnBeforeDelete;
-
+            
             Harmony.PatchAll(Assembly.GetExecutingAssembly());
             Log.LogInfo("Plugin is patched.");
         }